Minor fixes for Android
git-svn-id: svn://svn.code.sf.net/p/libkqueue/code/trunk@593 fb4e3144-bc1c-4b72-a658-5bcd248dd7f7

+++ src/linux/vnode.c
@@ -77,7 +77,7 @@ get_one_event(struct inotify_event *dst, int inofd)
             break;
         }
     }
-    dbg_printf("read(2) from inotify wd: %zu bytes", n);
+    dbg_printf("read(2) from inotify wd: %ld bytes", (long) n);
 
     /* FIXME-TODO: if len > 0, read(len) */
     if (dst->len != 0)
